baseline_requirements.txt
evotorch==0.5.1
swig==4.2.1.post0
gymnasium==1.0.0
pygame==2.6.1
#Pygments==2.18.0
#PyOpenGL==3.1.7
#moviepy==2.1.1
#imageio==2.36.0
#imageio-ffmpeg==0.5.1
box2d-py==2.3.5
cma==4.0.0
mujoco==3.2.5
#scikit-learn==1.5.2
#scipy==1.14.1